+++ persistent/ChangeLog.md
@@ -15,8 +15,10 @@
 * [#1063](https://github.com/yesodweb/persistent/pull/1063)
   * A new class member `keyFromRecordM` allows you to construct a `Key
     record` from a `record` if it was defined with `Primary`.
-* [#1036](https://github.com/yesodweb/persistent/pull/1036):
+* [#1036](https://github.com/yesodweb/persistent/pull/1036)
   * The method `entityIdFromJSON` that is used to parse entities now correctly works for entities that define a custom `Primary` key.
+* [#856](https://github.com/yesodweb/persistent/pull/856)
+  * Modify `upsertBy` to use backend-specific implementation (if any).
 * [#1066](https://github.com/yesodweb/persistent/pull/1066)
   * You can set a column's `sql=id` for a non `Id` column.
 * Fix a bug where unsafe migration error messages were being shown using `Show` prior to printing, resulting in less helpful output. [#1080](https://github.com/yesodweb/persistent/pull/1080)